backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
fastapi==0.115.6
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.12.0
future==1.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.0
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
